fix: fixed components order and names
