+++ src/app/home/components/JoinOurCommunity.tsx
@@ -1,53 +1,83 @@
 'use client'
-import { PrimaryButton } from '@/core/PrimaryButton'
+import { useEffect, useRef, useState } from 'react'
 import { images } from '@/data/images_url'
-import useIsLargeScreen from '@/hooks'
-import Link from 'next/link'
+
+const { imagesData } = images
 
 export const JoinOurCommunity = () => {
-  const isLargeScreen = useIsLargeScreen(1300)
+  const videoRef = useRef<HTMLVideoElement>(null)
+  const [showVideo, setShowVideo] = useState(false)
+  const [currentIndex, setCurrentIndex] = useState(0)
+  const [images] = useState(imagesData)
 
-  const joinUsImage = isLargeScreen
-    ? images.joinSectionDesktop
-    : images.joinSectionMobile
+  useEffect(() => {
+    if (videoRef.current) {
+      videoRef.current.volume = 0.5 // Establece el volumen al 50%
+    }
+  }, [])
+
+  const handlePlayVideo = () => {
+    setShowVideo(true)
+  }
 
   return (
-    <div className='relative m-auto flex w-4/6 max-w-[1920px] flex-col items-center justify-center py-2 lg:mb-32'>
-      <div className='z-0 flex flex-col items-center overflow-hidden rounded-xl bg-black-4 shadow-xl md:flex-row-reverse'>
-        <div className='flex h-full'>
-          <img
-            alt=''
-            className='flex h-full w-full'
-            height={600}
-            src={joinUsImage}
-            width={600}
-          />
+    <>
+      <h2 className='mb-8 ml-28 mt-[100px] text-3xl font-bold text-[#082965] md:text-4xl'>
+        ¿Por qué unirse a nosotros?
+      </h2>
+      <div className='relative m-auto mb-[200px] grid grid-cols-1 grid-rows-2 flex-col justify-center font-darker-grotesque md:max-w-[1980px] md:grid-cols-[1fr_1.35fr] md:grid-rows-1'>
+        <div
+          className={
+            'relative flex flex-col items-center justify-center bg-[#082965] md:items-start md:bg-blend-normal lg:pl-24'
+          }
+        >
+          <p className='text-center text-4 font-semibold leading-4 text-[#FCFCFC] md:w-[70%] md:text-start md:text-5 lg:text-11 lg:leading-[1.4]'>
+            Conéctate con profesionales ágiles de toda Latinoamérica. Accede
+            recursos exclusivos y participa en eventos que impulsarán tu
+            crecimiento.
+          </p>
+          <p className='text-center text-4 font-semibold text-[#FFFFFF] md:w-[70%] md:text-start md:text-5 lg:mt-8 lg:text-11 lg:leading-[1.4]'>
+            ¡<a className='text-[#FE5833]'>Regístrate </a> hoy y{' '}
+            <label className='text-[#FE5833]'>sé parte</label> de nuestra
+            transformación ágil!
+          </p>
         </div>
-        <div className='m-0 flex flex-col items-center gap-[2vw] bg-transparent pb-10 pt-[2rem] sm:w-1/2 md:items-start md:pb-0 md:pl-14 lg:justify-center lg:gap-0 lg:py-8'>
-          <p className='m-0 text-center font-darker-grotesque text-[5dvw] font-extrabold leading-8 text-blue-6 md:mb-0 md:flex md:flex-col md:text-start lg:text-15'>
-            ¿POR QUÉ UNIRSE
-          </p>
-          <p className='m-0 text-center font-darker-grotesque text-[6dvw] font-extrabold leading-8 text-red-400 md:mt-2 md:flex md:flex-col md:text-start md:text-17 lg:text-17'>
-            A NOSOTROS?
-          </p>
-          <p className='w-[80%] text-center font-karla text-[3.5dvw] font-medium leading-3 text-blue-6 md:flex md:text-start md:text-7 md:leading-9 lg:pt-10'>
-            Conéctate con profesionales ágiles de toda Latinoamérica, accede a
-            recursos exclusivos, y participa en eventos y webinars que
-            impulsarán tu crecimiento.{' '}
-          </p>
-          <p className='w-[80%] text-center font-karla font-bold leading-3 text-blue-6 md:text-start md:text-7 md:leading-9'>
-            ¡Regístrate hoy y sé parte de nuestra transformación ágil!
-          </p>
-          <Link href='/benefits'>
-            <div className='mt-3 flex items-center justify-center'>
-              <PrimaryButton
-                className='rounded-[20px] bg-red-400 px-2 py-[.5vw] font-darker-grotesque md:px-5'
-                label='Beneficios'
+        <div className='overflow-hidden bg-white shadow-lg'>
+          <div className='bg-black relative aspect-video'>
+            {!showVideo ? (
+              <div className='absolute inset-0 flex items-center justify-center'>
+                <img
+                  src='https://appwiseinnovations.dev/scrumlatam/Captura%20de%20pantalla%202025-07-06%20192049.png'
+                  alt='Video thumbnail'
+                  className='h-full w-full object-cover'
+                />
+                <button
+                  onClick={handlePlayVideo}
+                  className='bg-black absolute inset-0 flex items-center justify-center bg-opacity-30 transition-all hover:bg-opacity-40'
+                >
+                  <div className='flex h-16 w-16 items-center justify-center shadow-lg'>
+                    <svg
+                      className='h-30 w-30 ml-1 text-[#bcb7b7]'
+                      fill='currentColor'
+                      viewBox='0 0 24 24'
+                    >
+                      <path d='M8 5v14l11-7z' />
+                    </svg>
+                  </div>
+                </button>
+              </div>
+            ) : (
+              <video
+                ref={videoRef}
+                src='https://appwiseinnovations.dev/scrumlatam/video_otro.mp4'
+                controls
+                autoPlay
+                className='h-full w-full'
               />
-            </div>
-          </Link>
+            )}
+          </div>
         </div>
       </div>
-    </div>
+    </>
   )
 }
